LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/sameDocument-after-navigate-restore-expected.txt



FAIL entry.sameDocument is properly restored after cross-document back assert_equals: expected 2 but got 0
